+++ testing/helpers/scheduler/contentReadyEvent.js
