+++ testing/tests/DevExpress.ui.widgets.scheduler.integrationBase/integration.appointments.tests.js
@@ -14,7 +14,7 @@ import { DataSource } from 'data/data_source/data_source';
 import CustomStore from 'data/custom_store';
 import dataUtils from 'core/element_data';
 import dateSerialization from 'core/utils/date_serialization';
-import { SchedulerTestWrapper, initTestMarkup, createWrapper, CLASSES } from './helpers.js';
+import { SchedulerTestWrapper, initTestMarkup, createWrapper, CLASSES } from '../../helpers/scheduler/helpers.js';
 import browser from 'core/utils/browser';
 import { Deferred } from 'core/utils/deferred';
