+++ testing/tests/DevExpress.ui.widgets.scheduler.integrationBase/integration.appointmentsWithTimezone.tests.js
@@ -9,7 +9,7 @@ import { DataSource } from 'data/data_source/data_source';
 import subscribes from 'ui/scheduler/ui.scheduler.subscribes';
 import dataUtils from 'core/element_data';
 import dateUtils from 'core/utils/date';
-import { SchedulerTestWrapper } from './helpers.js';
+import { SchedulerTestWrapper } from '../../helpers/scheduler/helpers.js';
 
 import 'common.css!';
 import 'generic_light.css!';
