+++ testing/tests/DevExpress.ui.widgets.scheduler.integrationBase/integration.dstAppointments.tests.js
@@ -1,4 +1,4 @@
-import { initTestMarkup, createWrapper } from './helpers.js';
+import { initTestMarkup, createWrapper } from '../../helpers/scheduler/helpers.js';
 import dateLocalization from 'localization/date';
 import fx from 'animation/fx';
 import subscribes from 'ui/scheduler/ui.scheduler.subscribes';
