+++ testing/tests/DevExpress.ui.widgets.scheduler.integrationBase/integration.recurrenceRuleValidation.tests.js
