+++ testing/tests/DevExpress.ui.widgets.scheduler/dragAndDropAppointments.tests.js
@@ -7,7 +7,7 @@ import {
     createWrapper,
     initTestMarkup,
     isDesktopEnvironment
-} from './helpers.js';
+} from '../../helpers/scheduler/helpers.js';
 
 import 'common.css!';
 import 'generic_light.css!';
